packages/council/pyproject.toml
[project]
name = "council"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Dhyan S Agni", email = "dhyanagni19@gmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "llm-gateway",
    "python-dotenv>=1.2.2",
]

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"

[tool.uv.sources]
llm-gateway = { workspace = true }
